requirements.txt
web3>=6.0.0
paho-mqtt>=1.6.1
py-solc-x>=1.1.1